Rename files and pkg according to Go conventions
