+++ node/cmd/txverifier/testdata/fuzz/FuzzParseLogMessagePublishedPayload/3360e0946e7f34d0
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("0000000000000000000000000000000000000000000000000000000000000000000")
